Expose QR code generation functions.
